clean struct
